requirements.txt
Flask==0.12.2
Flask-API==1.0
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
numpy==1.13.3
pillow==4.2.1
keras==2.1.5
tensorflow==1.7.0
h5py==2.7.0
pytesseract==0.2.0

